app/requirements.txt
boto3
psycopg2-binary
Flask
flask-cors
python-dotenv
awsgi
pytest
werkzeug
PyJWT